% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_cfar.r
\name{g_cfar1}
\alias{g_cfar1}
\title{CFAR program
based on the original codes of Xialu Liu of SDSU.}
\usage{
g_cfar1(tmax = 1001, rho = 5, phi_func = NULL, grid = 1000, sigma = 1)
}
\description{
CFAR program
based on the original codes of Xialu Liu of SDSU.
}
