% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.r
\name{p_cfar_part}
\alias{p_cfar_part}
\title{partial prediction
Assume that we have t curves and want to predit the curve at time t+1, but we know the first n observations in the curve, to predict the n+1 observation.}
\usage{
p_cfar_part(model, f, new.obs)
}
\description{
partial prediction
Assume that we have t curves and want to predit the curve at time t+1, but we know the first n observations in the curve, to predict the n+1 observation.
}
